Extract pipeline batch and models orchestration
