+++ src/cja_auto_sdr/generator.py
@@ -334,173 +334,7 @@ RUN_SUMMARY_SCHEMA_VERSION = "1.1"
 # ==================== DATA STRUCTURES ====================
 
 
-@dataclass
-class ProcessingResult:
-    """Result of processing a single data view"""
-
-    data_view_id: str
-    data_view_name: str
-    success: bool
-    duration: float
-    metrics_count: int = 0
-    dimensions_count: int = 0
-    dq_issues_count: int = 0
-    dq_issues: list[ValidationIssue] = field(default_factory=list)
-    dq_severity_counts: dict[str, int] = field(default_factory=dict)
-    output_file: str = ""
-    output_files: list[str] = field(default_factory=list)
-    error_message: str = ""
-    failure_code: str = ""
-    failure_reason: str = ""
-    partial_output: bool = False
-    partial_reasons: list[str] = field(default_factory=list)
-    file_size_bytes: int = 0
-    # Inventory statistics (populated when inventory options are used)
-    segments_count: int = 0
-    segments_high_complexity: int = 0
-    calculated_metrics_count: int = 0
-    calculated_metrics_high_complexity: int = 0
-    derived_fields_count: int = 0
-    derived_fields_high_complexity: int = 0
-
-    def __post_init__(self) -> None:
-        """Normalize output and partial-run observability fields for all result producers."""
-        self.output_file, self.output_files = _normalize_output_artifact_state(
-            self.output_file,
-            self.output_files,
-        )
-        self.partial_output, self.partial_reasons = _normalize_partial_state(
-            self.partial_output,
-            self.partial_reasons,
-        )
-
-    @property
-    def file_size_formatted(self) -> str:
-        """Return human-readable file size (e.g., '1.5 MB', '256 KB')."""
-        return format_file_size(self.file_size_bytes)
-
-    @property
-    def emitted_output_files(self) -> list[str]:
-        """Return the normalized emitted artifact list."""
-        return list(self.output_files)
-
-    @property
-    def has_inventory(self) -> bool:
-        """Check if any inventory data was collected."""
-        return self.segments_count > 0 or self.calculated_metrics_count > 0 or self.derived_fields_count > 0
-
-    @property
-    def total_high_complexity(self) -> int:
-        """Total count of high-complexity items across all inventories."""
-        return (
-            self.segments_high_complexity
-            + self.calculated_metrics_high_complexity
-            + self.derived_fields_high_complexity
-        )
-
-
-@dataclass
-class WorkerArgs:
-    """Arguments for process_single_dataview_worker (replaces opaque tuple)."""
-
-    data_view_id: str
-    config_file: str = "config.json"
-    output_dir: str = "."
-    log_level: str = "INFO"
-    log_format: str = "text"
-    output_format: str = "excel"
-    enable_cache: bool = False
-    cache_size: int = 1000
-    cache_ttl: int = 3600
-    quiet: bool = False
-    skip_validation: bool = False
-    max_issues: int = 0
-    clear_cache: bool = False
-    show_timings: bool = False
-    metrics_only: bool = False
-    dimensions_only: bool = False
-    profile: str | None = None
-    shared_cache: SharedValidationCache | None = None
-    api_tuning_config: APITuningConfig | None = None
-    circuit_breaker_config: CircuitBreakerConfig | None = None
-    include_derived_inventory: bool = False
-    include_calculated_metrics: bool = False
-    include_segments_inventory: bool = False
-    inventory_only: bool = False
-    inventory_order: list[str] | None = None
-    quality_report_only: bool = False
-    allow_partial: bool = False
-    production_mode: bool = False
-    batch_id: str | None = None
-
-
-@dataclass(frozen=True)
-class ProcessingConfig:
-    """Configuration bundle for single data view processing."""
-
-    config_file: str = "config.json"
-    output_dir: str | Path = "."
-    log_level: str = "INFO"
-    log_format: str = "text"
-    output_format: str = "excel"
-    enable_cache: bool = False
-    cache_size: int = 1000
-    cache_ttl: int = 3600
-    quiet: bool = False
-    skip_validation: bool = False
-    max_issues: int = 0
-    clear_cache: bool = False
-    show_timings: bool = False
-    metrics_only: bool = False
-    dimensions_only: bool = False
-    profile: str | None = None
-    shared_cache: SharedValidationCache | None = None
-    api_tuning_config: APITuningConfig | None = None
-    circuit_breaker_config: CircuitBreakerConfig | None = None
-    include_derived_inventory: bool = False
-    include_calculated_metrics: bool = False
-    include_segments_inventory: bool = False
-    inventory_only: bool = False
-    inventory_order: list[str] | None = None
-    quality_report_only: bool = False
-    allow_partial: bool = False
-    production_mode: bool = False
-    batch_id: str | None = None
-
-
-@dataclass(frozen=True)
-class BatchConfig:
-    """Configuration bundle for batch processing."""
-
-    config_file: str = "config.json"
-    output_dir: str = "."
-    workers: int = 4
-    continue_on_error: bool = False
-    log_level: str = "INFO"
-    log_format: str = "text"
-    output_format: str = "excel"
-    enable_cache: bool = False
-    cache_size: int = 1000
-    cache_ttl: int = 3600
-    quiet: bool = False
-    skip_validation: bool = False
-    max_issues: int = 0
-    clear_cache: bool = False
-    show_timings: bool = False
-    metrics_only: bool = False
-    dimensions_only: bool = False
-    profile: str | None = None
-    shared_cache: bool = False
-    api_tuning_config: APITuningConfig | None = None
-    circuit_breaker_config: CircuitBreakerConfig | None = None
-    include_derived_inventory: bool = False
-    include_calculated_metrics: bool = False
-    include_segments_inventory: bool = False
-    inventory_only: bool = False
-    inventory_order: list[str] | None = None
-    quality_report_only: bool = False
-    allow_partial: bool = False
-    production_mode: bool = False
+from cja_auto_sdr.pipeline.models import BatchConfig, ProcessingConfig, ProcessingResult, WorkerArgs
 
 
 @dataclass(frozen=True)
@@ -5594,438 +5428,16 @@ def process_single_dataview(
 
 
 def process_single_dataview_worker(args: WorkerArgs) -> ProcessingResult:
-    """Worker function for multiprocessing.
+    """Worker function for multiprocessing."""
+    from cja_auto_sdr.pipeline.workers import process_single_dataview_worker as _impl
 
-    Args:
-        args: A WorkerArgs dataclass with all processing parameters.
-
-    Returns:
-        ProcessingResult
-    """
-    # Retry config is propagated via env vars and resolved at call time
-    # by _effective_retry_config() in resilience.py — no global mutation needed.
-
-    return process_single_dataview(
-        args.data_view_id,
-        processing_config=ProcessingConfig(
-            config_file=args.config_file,
-            output_dir=args.output_dir,
-            log_level=args.log_level,
-            log_format=args.log_format,
-            output_format=args.output_format,
-            enable_cache=args.enable_cache,
-            cache_size=args.cache_size,
-            cache_ttl=args.cache_ttl,
-            quiet=args.quiet,
-            skip_validation=args.skip_validation,
-            max_issues=args.max_issues,
-            clear_cache=args.clear_cache,
-            show_timings=args.show_timings,
-            metrics_only=args.metrics_only,
-            dimensions_only=args.dimensions_only,
-            profile=args.profile,
-            shared_cache=args.shared_cache,
-            api_tuning_config=args.api_tuning_config,
-            circuit_breaker_config=args.circuit_breaker_config,
-            include_derived_inventory=args.include_derived_inventory,
-            include_calculated_metrics=args.include_calculated_metrics,
-            include_segments_inventory=args.include_segments_inventory,
-            inventory_only=args.inventory_only,
-            inventory_order=args.inventory_order,
-            quality_report_only=args.quality_report_only,
-            allow_partial=args.allow_partial,
-            production_mode=args.production_mode,
-            batch_id=args.batch_id,
-        ),
-    )
+    return _impl(args)
 
 
 # ==================== BATCH PROCESSOR CLASS ====================
 
 
-class BatchProcessor:
-    """
-    Process multiple data views in parallel using multiprocessing.
-
-    Provides parallel execution of SDR generation across multiple data views
-    with configurable worker count and error handling.
-
-    Args:
-        config_file: Path to CJA config file (default: 'config.json')
-        output_dir: Directory for output files (default: current directory)
-        workers: Number of parallel workers, 1-256 (default: 4)
-        continue_on_error: Continue if individual data views fail (default: False)
-        log_level: Logging level - DEBUG, INFO, WARNING, ERROR, CRITICAL (default: INFO)
-        log_format: Log output format - "text" (default) or "json" for structured logging
-        output_format: Output format - excel, csv, json, html, markdown, all (default: excel)
-        enable_cache: Enable validation result caching (default: False)
-        cache_size: Maximum cached validation results, >= 1 (default: 1000)
-        cache_ttl: Cache time-to-live in seconds, >= 1 (default: 3600)
-        quiet: Suppress non-error output (default: False)
-        skip_validation: Skip data quality validation (default: False)
-        max_issues: Limit issues to top N by severity, >= 0; 0 = all (default: 0)
-        clear_cache: Clear validation cache before processing (default: False)
-        show_timings: Display performance timing for each data view (default: False)
-        metrics_only: Only include metrics in output (default: False)
-        dimensions_only: Only include dimensions in output (default: False)
-        profile: Profile name for credentials (default: None)
-        shared_cache: Share validation cache across batch workers (default: False)
-        api_tuning_config: API worker auto-tuning configuration (default: None)
-        circuit_breaker_config: Circuit breaker configuration (default: None)
-        production_mode: Reduce per-issue warning log volume for production runs (default: False)
-    """
-
-    def __init__(
-        self,
-        config_file: str = "config.json",
-        output_dir: str = ".",
-        workers: int = 4,
-        continue_on_error: bool = False,
-        log_level: str = "INFO",
-        log_format: str = "text",
-        output_format: str = "excel",
-        enable_cache: bool = False,
-        cache_size: int = 1000,
-        cache_ttl: int = 3600,
-        quiet: bool = False,
-        skip_validation: bool = False,
-        max_issues: int = 0,
-        clear_cache: bool = False,
-        show_timings: bool = False,
-        metrics_only: bool = False,
-        dimensions_only: bool = False,
-        profile: str | None = None,
-        shared_cache: bool = False,
-        api_tuning_config: APITuningConfig | None = None,
-        circuit_breaker_config: CircuitBreakerConfig | None = None,
-        include_derived_inventory: bool = False,
-        include_calculated_metrics: bool = False,
-        include_segments_inventory: bool = False,
-        inventory_only: bool = False,
-        inventory_order: list[str] | None = None,
-        quality_report_only: bool = False,
-        allow_partial: bool = False,
-        production_mode: bool = False,
-        batch_config: BatchConfig | None = None,
-    ):
-        if batch_config is None:
-            batch_config = BatchConfig(
-                config_file=config_file,
-                output_dir=output_dir,
-                workers=workers,
-                continue_on_error=continue_on_error,
-                log_level=log_level,
-                log_format=log_format,
-                output_format=output_format,
-                enable_cache=enable_cache,
-                cache_size=cache_size,
-                cache_ttl=cache_ttl,
-                quiet=quiet,
-                skip_validation=skip_validation,
-                max_issues=max_issues,
-                clear_cache=clear_cache,
-                show_timings=show_timings,
-                metrics_only=metrics_only,
-                dimensions_only=dimensions_only,
-                profile=profile,
-                shared_cache=shared_cache,
-                api_tuning_config=api_tuning_config,
-                circuit_breaker_config=circuit_breaker_config,
-                include_derived_inventory=include_derived_inventory,
-                include_calculated_metrics=include_calculated_metrics,
-                include_segments_inventory=include_segments_inventory,
-                inventory_only=inventory_only,
-                inventory_order=inventory_order,
-                quality_report_only=quality_report_only,
-                allow_partial=allow_partial,
-                production_mode=production_mode,
-            )
-
-        self.config_file = batch_config.config_file
-        self.output_dir = batch_config.output_dir
-        self.clear_cache = batch_config.clear_cache
-        self.show_timings = batch_config.show_timings
-        self.metrics_only = batch_config.metrics_only
-        self.dimensions_only = batch_config.dimensions_only
-        self.workers = batch_config.workers
-        self.continue_on_error = batch_config.continue_on_error
-        self.log_level = batch_config.log_level
-        self.log_format = batch_config.log_format
-        self.output_format = batch_config.output_format
-        self.enable_cache = batch_config.enable_cache
-        self.cache_size = batch_config.cache_size
-        self.cache_ttl = batch_config.cache_ttl
-        self.quiet = batch_config.quiet
-        self.skip_validation = batch_config.skip_validation
-        self.max_issues = batch_config.max_issues
-        self.profile = batch_config.profile
-        self.shared_cache_enabled = batch_config.shared_cache
-        self.api_tuning_config = batch_config.api_tuning_config
-        self.circuit_breaker_config = batch_config.circuit_breaker_config
-        self.include_derived_inventory = batch_config.include_derived_inventory
-        self.include_calculated_metrics = batch_config.include_calculated_metrics
-        self.include_segments_inventory = batch_config.include_segments_inventory
-        self.inventory_only = batch_config.inventory_only
-        self.inventory_order = batch_config.inventory_order
-        self.quality_report_only = batch_config.quality_report_only
-        self.allow_partial = batch_config.allow_partial
-        self.production_mode = batch_config.production_mode
-        self.batch_id = str(uuid.uuid4())[:8]  # Short correlation ID for log tracing
-        base_logger = setup_logging(batch_mode=True, log_level=self.log_level, log_format=self.log_format)
-        self.logger = with_log_context(base_logger, run_mode="batch", batch_id=self.batch_id)
-        self.logger.info(f"Batch ID: {self.batch_id}")
-
-        # Create shared validation cache if enabled
-        self._shared_cache: SharedValidationCache | None = None
-        if self.shared_cache_enabled and self.enable_cache and not self.skip_validation:
-            self._shared_cache = SharedValidationCache(max_size=self.cache_size, ttl_seconds=self.cache_ttl)
-            self.logger.info(f"[{self.batch_id}] Shared validation cache enabled (max_size={self.cache_size})")
-
-        # Create output directory if it doesn't exist
-        try:
-            Path(self.output_dir).mkdir(parents=True, exist_ok=True)
-        except PermissionError as e:
-            raise OutputError(
-                f"Permission denied creating output directory: {self.output_dir}. "
-                "Check that you have write permissions for the parent directory.",
-            ) from e
-        except OSError as e:
-            raise OutputError(
-                f"Cannot create output directory '{self.output_dir}': {e}. "
-                "Verify the path is valid and the disk has available space.",
-            ) from e
-
-    def process_batch(self, data_view_ids: list[str]) -> dict:
-        """
-        Process multiple data views in parallel
-
-        Args:
-            data_view_ids: List of data view IDs to process
-
-        Returns:
-            Dictionary with processing results
-        """
-        self.logger.info("=" * BANNER_WIDTH)
-        self.logger.info(f"[{self.batch_id}] BATCH PROCESSING START")
-        self.logger.info("=" * BANNER_WIDTH)
-        self.logger.info(f"[{self.batch_id}] Data views to process: {len(data_view_ids)}")
-        self.logger.info(f"[{self.batch_id}] Parallel workers: {self.workers}")
-        self.logger.info(f"[{self.batch_id}] Continue on error: {self.continue_on_error}")
-        self.logger.info(f"[{self.batch_id}] Output directory: {self.output_dir}")
-        self.logger.info(f"[{self.batch_id}] Output format: {self.output_format}")
-        self.logger.info("=" * BANNER_WIDTH)
-
-        batch_start_time = time.time()
-
-        results = {"successful": [], "failed": [], "total": len(data_view_ids), "total_duration": 0}
-
-        # Prepare arguments for each worker
-        worker_args = [
-            WorkerArgs(
-                data_view_id=dv_id,
-                config_file=self.config_file,
-                output_dir=self.output_dir,
-                log_level=self.log_level,
-                log_format=self.log_format,
-                output_format=self.output_format,
-                enable_cache=self.enable_cache,
-                cache_size=self.cache_size,
-                cache_ttl=self.cache_ttl,
-                quiet=self.quiet,
-                skip_validation=self.skip_validation,
-                max_issues=self.max_issues,
-                clear_cache=self.clear_cache,
-                show_timings=self.show_timings,
-                metrics_only=self.metrics_only,
-                dimensions_only=self.dimensions_only,
-                profile=self.profile,
-                shared_cache=self._shared_cache,
-                api_tuning_config=self.api_tuning_config,
-                circuit_breaker_config=self.circuit_breaker_config,
-                include_derived_inventory=self.include_derived_inventory,
-                include_calculated_metrics=self.include_calculated_metrics,
-                include_segments_inventory=self.include_segments_inventory,
-                inventory_only=self.inventory_only,
-                inventory_order=self.inventory_order,
-                quality_report_only=self.quality_report_only,
-                allow_partial=self.allow_partial,
-                production_mode=self.production_mode,
-                batch_id=self.batch_id,
-            )
-            for dv_id in data_view_ids
-        ]
-
-        try:
-            # Process with ProcessPoolExecutor for true parallelism
-            with ProcessPoolExecutor(max_workers=self.workers) as executor:
-                # Submit all tasks
-                future_to_dv = {
-                    executor.submit(process_single_dataview_worker, wa): wa.data_view_id for wa in worker_args
-                }
-
-                # Collect results as they complete with progress bar
-                with tqdm(
-                    total=len(data_view_ids),
-                    desc="Processing data views",
-                    unit="view",
-                    bar_format="{l_bar}{bar}| {n_fmt}/{total_fmt} [{elapsed}<{remaining}, {rate_fmt}]",
-                    disable=self.quiet,
-                ) as pbar:
-                    for future in as_completed(future_to_dv):
-                        dv_id = future_to_dv[future]
-                        try:
-                            result = future.result()
-
-                            if result.success:
-                                results["successful"].append(result)
-                                pbar.set_postfix_str(f"✓ {dv_id[:20]}", refresh=True)
-                                self.logger.info(f"[{self.batch_id}] ✓ {dv_id}: SUCCESS ({result.duration:.1f}s)")
-                            else:
-                                results["failed"].append(result)
-                                pbar.set_postfix_str(f"✗ {dv_id[:20]}", refresh=True)
-                                self.logger.error(f"[{self.batch_id}] ✗ {dv_id}: FAILED - {result.error_message}")
-
-                                if not self.continue_on_error:
-                                    self.logger.warning(
-                                        f"[{self.batch_id}] Stopping batch processing due to error (use --continue-on-error to continue)",
-                                    )
-                                    # Cancel remaining tasks
-                                    for f in future_to_dv:
-                                        f.cancel()
-                                    break
-
-                        except KeyboardInterrupt, SystemExit:
-                            # Allow graceful shutdown on Ctrl+C
-                            self.logger.warning(f"[{self.batch_id}] Interrupted - cancelling remaining tasks...")
-                            for f in future_to_dv:
-                                f.cancel()
-                            raise
-                        except Exception as e:  # Intentional: batch worker resilience boundary
-                            is_expected = isinstance(e, RECOVERABLE_BATCH_WORKER_EXCEPTIONS)
-                            prefix = "EXCEPTION" if is_expected else "UNEXPECTED EXCEPTION"
-                            self.logger.error(f"[{self.batch_id}] ✗ {dv_id}: {prefix} - {e!s}")
-                            if not is_expected:
-                                self.logger.debug("Unexpected batch worker error", exc_info=True)
-                            results["failed"].append(
-                                ProcessingResult(
-                                    data_view_id=dv_id,
-                                    data_view_name="Unknown",
-                                    success=False,
-                                    duration=0,
-                                    error_message=str(e),
-                                    failure_code=FAILURE_CODE_BATCH_WORKER_EXCEPTION,
-                                    failure_reason=f"batch_worker_exception:{type(e).__name__}",
-                                ),
-                            )
-
-                            if not self.continue_on_error:
-                                self.logger.warning(f"[{self.batch_id}] Stopping batch processing due to error")
-                                for f in future_to_dv:
-                                    f.cancel()
-                                break
-
-                        pbar.update(1)
-        finally:
-            # Log shared cache statistics and always cleanup resources.
-            if self._shared_cache is not None:
-                cache_stats = self._shared_cache.get_statistics()
-                self.logger.info(
-                    f"[{self.batch_id}] Shared cache stats: {cache_stats['hits']} hits, "
-                    f"{cache_stats['misses']} misses ({cache_stats['hit_rate']:.1f}% hit rate)",
-                )
-                self._shared_cache.shutdown()
-                self._shared_cache = None
-
-        results["total_duration"] = time.time() - batch_start_time
-
-        # Print summary
-        self.print_summary(results)
-
-        return results
-
-    def print_summary(self, results: dict):
-        """Print detailed batch processing summary with color-coded output"""
-        total = results["total"]
-        successful_count = len(results["successful"])
-        failed_count = len(results["failed"])
-        success_rate = (successful_count / total * 100) if total > 0 else 0
-        total_duration = results["total_duration"]
-        avg_duration = (total_duration / total) if total > 0 else 0
-
-        # Calculate total output size
-        total_file_size = sum(r.file_size_bytes for r in results["successful"])
-        total_size_formatted = format_file_size(total_file_size)
-
-        # Log to file
-        self.logger.info("")
-        self.logger.info("=" * BANNER_WIDTH)
-        self.logger.info(f"[{self.batch_id}] BATCH PROCESSING SUMMARY")
-        self.logger.info("=" * BANNER_WIDTH)
-        self.logger.info(f"[{self.batch_id}] Total data views: {total}")
-        self.logger.info(f"[{self.batch_id}] Successful: {successful_count}")
-        self.logger.info(f"[{self.batch_id}] Failed: {failed_count}")
-        self.logger.info(f"[{self.batch_id}] Success rate: {success_rate:.1f}%")
-        self.logger.info(f"[{self.batch_id}] Total output size: {total_size_formatted}")
-        self.logger.info(f"[{self.batch_id}] Total duration: {total_duration:.1f}s")
-        self.logger.info(f"[{self.batch_id}] Average per data view: {avg_duration:.1f}s")
-        if total_duration > 0:
-            throughput = total / total_duration
-            self.logger.info(f"[{self.batch_id}] Throughput: {throughput:.2f} views/second")
-        self.logger.info("=" * BANNER_WIDTH)
-
-        # Print color-coded console output
-        print()
-        print("=" * BANNER_WIDTH)
-        print(ConsoleColors.bold("BATCH PROCESSING SUMMARY"))
-        print("=" * BANNER_WIDTH)
-        print(f"Total data views: {total}")
-        print(f"Successful: {ConsoleColors.success(str(successful_count))}")
-        if failed_count > 0:
-            print(f"Failed: {ConsoleColors.error(str(failed_count))}")
-        else:
-            print(f"Failed: {failed_count}")
-        print(f"Success rate: {ConsoleColors.status(success_rate == 100, f'{success_rate:.1f}%')}")
-        print(f"Total output size: {total_size_formatted}")
-        print(f"Total duration: {total_duration:.1f}s")
-        print(f"Average per data view: {avg_duration:.1f}s")
-        if total_duration > 0:
-            throughput = total / total_duration
-            print(f"Throughput: {throughput:.2f} views/second")
-        print()
-
-        if results["successful"]:
-            print(ConsoleColors.success("Successful Data Views:"))
-            self.logger.info("")
-            self.logger.info("Successful Data Views:")
-            for result in results["successful"]:
-                size_str = result.file_size_formatted
-                line = f"  {result.data_view_id:20s}  {result.data_view_name:30s}  {size_str:>10s}  {result.duration:5.1f}s"
-                print(ConsoleColors.success("  ✓") + line[3:])
-                self.logger.info(
-                    f"  ✓ {result.data_view_id:20s}  {result.data_view_name:30s}  {size_str:>10s}  {result.duration:5.1f}s",
-                )
-            print()
-            self.logger.info("")
-
-        if results["failed"]:
-            print(ConsoleColors.error("Failed Data Views:"))
-            self.logger.info("Failed Data Views:")
-            for result in results["failed"]:
-                line = f"  {result.data_view_id:20s}  {result.error_message}"
-                print(ConsoleColors.error("  ✗") + line[3:])
-                self.logger.info(f"  ✗ {result.data_view_id:20s}  {result.error_message}")
-            print()
-            self.logger.info("")
-
-        print("=" * BANNER_WIDTH)
-        self.logger.info("=" * BANNER_WIDTH)
-
-        if total > 0 and total_duration > 0:
-            throughput = (total / total_duration) * 60  # per minute
-            print(f"Throughput: {throughput:.1f} data views per minute")
-            print("=" * BANNER_WIDTH)
-            self.logger.info(f"Throughput: {throughput:.1f} data views per minute")
-            self.logger.info("=" * BANNER_WIDTH)
+BatchProcessor = importlib.import_module("cja_auto_sdr.pipeline.batch").BatchProcessor
 
 
 # ==================== DRY-RUN MODE ====================
