+++ src/cja_auto_sdr/pipeline/dry_run.py
@@ -1,7 +1,18 @@
-"""Dry-run mode (current implementation lives in generator)."""
+"""Dry-run wrapper."""
+
+from __future__ import annotations
+
+import logging
 
 __all__ = ["run_dry_run"]
 
-from cja_auto_sdr.core.lazy import make_getattr
 
-__getattr__ = make_getattr(__name__, __all__, target_module="cja_auto_sdr.generator")
+def _generator_module():
+    from cja_auto_sdr import generator as _generator
+
+    return _generator
+
+
+def run_dry_run(data_views: list[str], config_file: str, logger: logging.Logger, profile: str | None = None) -> bool:
+    """Delegate dry-run validation to the generator implementation."""
+    return _generator_module().run_dry_run(data_views, config_file, logger, profile=profile)
